+++ src/main/java/io/github/palexdev/mfxcore/observables/OnChanged.java
@@ -18,10 +18,13 @@
 
 package io.github.palexdev.mfxcore.observables;
 
+import io.github.palexdev.mfxcore.base.TriConsumer;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
 
+import java.lang.ref.WeakReference;
 import java.util.function.BiConsumer;
+import java.util.function.BiFunction;
 
 /**
  * Concrete implementation of {@link When} that uses {@link ChangeListener}s to
@@ -30,6 +33,9 @@ import java.util.function.BiConsumer;
  * You can specify the action to perform when this happens using a {@link BiConsumer},
  * {@link #then(BiConsumer)}.
  * <p>
+ * You can also set a condition that has to be met for the action to be executed (see {@link #condition(BiFunction)}),
+ * and an "else" action that is executed when it is not met, (see {@link #otherwise(TriConsumer)}).
+ * <p></p>
  * To activate the construct do not forget to call {@link #listen()} at the end.
  * <p></p>
  * An example:
@@ -37,7 +43,9 @@ import java.util.function.BiConsumer;
  * {@code
  *      IntegerProperty aNumber = new SimpleIntegerProperty(69);
  *      When.onChanged(aNumber) // You can also use... OnChanged.forObservable(...)
+ *              .condition(aCondition)
  *              .then((oldValue, newValue) -> System.out.println("Value switched from: " + oldValue + " to " + newValue))
+ *              .otherwise((ref, oldValue, newValue) -> System.out.println("Condition not met, execution action B"))
  *              .oneShot()
  *              .listen();
  * }
@@ -49,6 +57,9 @@ public class OnChanged<T> extends When<T> {
 	//================================================================================
 	private ChangeListener<T> listener;
 	private BiConsumer<T, T> action;
+	private TriConsumer<WeakReference<When<T>>, T, T> otherwise = (w, o, n) -> {
+	};
+	private BiFunction<T, T, Boolean> condition = (o, n) -> true;
 
 	//================================================================================
 	// Constructors
@@ -83,6 +94,46 @@ public class OnChanged<T> extends When<T> {
 	}
 
 	/**
+	 * Allows to set an action to perform when the given {@link #condition(BiFunction)} is not met.
+	 * <p></p>
+	 * This makes the "system" much more versatile. Imagine having a one-shot listener that you want to
+	 * dispose anyway even if the condition is not met, you can write something like this;
+	 * <pre>
+	 * {@code
+	 * When.onChanged(observable)
+	 *      .condition(aCondition)
+	 *      .then(action)
+	 *      .otherwise((w, o, n) -> Optional.ofNullable(w.get()).ifPresent(When::dispose)) // Note the null check
+	 *      .listen();
+	 *
+	 * }
+	 * </pre>
+	 * <p></p>
+	 * Also note that the otherwise action also carries the reference to this object wrapped in a {@link WeakReference}.
+	 */
+	public OnChanged<T> otherwise(TriConsumer<WeakReference<When<T>>, T, T> otherwise) {
+		this.otherwise = otherwise;
+		return this;
+	}
+
+	/**
+	 * Allows to specify a condition under which the set action (see {@link #then(BiConsumer)})
+	 * is to be executed.
+	 * <p></p>
+	 * The condition is specified through a {@link BiFunction} that provides both the old and new values
+	 * of the {@link ObservableValue}.
+	 * <p></p>
+	 * In case the condition is not met the {@link #otherwise(TriConsumer)} action is executed instead.
+	 * <p></p>
+	 * For one-shot listeners, the action is executed and the listener disposed only if the condition is met, else
+	 * the {@link #otherwise(TriConsumer)} action is executed instead.
+	 */
+	public OnChanged<T> condition(BiFunction<T, T, Boolean> condition) {
+		this.condition = condition;
+		return this;
+	}
+
+	/**
 	 * Activates the {@code OnChanged} construct with the previously specified parameters.
 	 * So, builds the {@link ChangeListener} according to the {@link #isOneShot()} parameter,
 	 * then adds the listener to the specified {@link ObservableValue} and finally puts the Observable and
@@ -92,11 +143,21 @@ public class OnChanged<T> extends When<T> {
 	public OnChanged<T> listen() {
 		if (oneShot) {
 			listener = (observable, oldValue, newValue) -> {
-				action.accept(oldValue, newValue);
-				dispose();
+				if (condition.apply(oldValue, newValue)) {
+					action.accept(oldValue, newValue);
+					dispose();
+				} else {
+					otherwise.accept(new WeakReference<>(this), oldValue, newValue);
+				}
 			};
 		} else {
-			listener = (observable, oldValue, newValue) -> action.accept(oldValue, newValue);
+			listener = (observable, oldValue, newValue) -> {
+				if (condition.apply(oldValue, newValue)) {
+					action.accept(oldValue, newValue);
+				} else {
+					otherwise.accept(new WeakReference<>(this), oldValue, newValue);
+				}
+			};
 		}
 
 		observableValue.addListener(listener);