+++ CHANGELOG.md
@@ -14,6 +14,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - **Removed** for now removed features.
 - **Fixed** for any bug fixes.
 
+## [11.0.7] - 18-07-2022
+
+## Added
+
+- Added "condition" and "otherwise" capabilities to OnChanged and OnInvalidated constructs
+- Added inHierarchy methods for GestureEvents to NodeUtils
+
 ## [11.0.6] - 20-06-2022
 
 ## Added